Add NeverEmptyRule & refactor function rules
