+++ CHANGELOG.md
@@ -5,6 +5,11 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+## [2.1.0] - 2025-10-13
+### Added
+- [GH#17](https://github.com/Korbeil/phpstan-generic-rules/pull/17) Add NeverEmptyRule
+- [GH#17](https://github.com/Korbeil/phpstan-generic-rules/pull/17) Refactor function rules
+
 ## [2.0.0] - 2024-11-25
 ### Added
 - [GH#14](https://github.com/Korbeil/phpstan-generic-rules/pull/14) PHPStan 2.0 support
